add api test
